@@ -1,11 +1,16 @@
+import dataclasses
 import logging
 import re
 import sys
+from dataclasses import Field
+from pprint import pformat
 from typing import List, Tuple
 
-from logictree.nodes import control, ops
+from logictree.nodes import LogicMux, control, ops
 from logictree.nodes.control.assign import LogicAssign
 from logictree.nodes.ops import LogicConst, LogicVar
+from logictree.nodes.ops.comparison import EqOp, NeqOp
+from logictree.nodes.ops.gates import AndOp, NotOp
 from logictree.nodes.selects import BitSelect, Concat, PartSelect
 from logictree.nodes.struct.module import Module
 from logictree.utils.display import pretty_print
@@ -21,6 +26,14 @@ _BINARY_RE = re.compile(
     r"^(?P<width>\d+)\s*'\s*(?P<base>[bBoOdDhH])\s*(?P<digits>[_0-9a-fA-FxzXZ]+)$"
 )
 _RANGE_RE = re.compile(r"\[\s*(\d+)\s*:\s*(\d+)\s*\]")
+
+
+def contains_field_object(obj):
+    for name, val in vars(obj).items():
+        if isinstance(val, Field):
+            return name, val
+    return None, None
+
 
 def _parse_sv_int_literal(txt: str) -> int:
     s = txt.strip().lower().replace("_", "")
@@ -44,12 +57,18 @@ def where_defined(obj):
     # inspect.getfile is helpful too, but can raise for some objects
     return path or f"(module {obj.__module__} has no __file__)"
 
-log.warning("Parser module: %s @ %s",
-            SystemVerilogSubsetParser.__module__,
-            where_defined(SystemVerilogSubsetParser))
-log.warning("Visitor module: %s @ %s",
-            SystemVerilogSubsetVisitor.__module__,
-            where_defined(SystemVerilogSubsetVisitor))
+
+log.warning(
+    "Parser module: %s @ %s",
+    SystemVerilogSubsetParser.__module__,
+    where_defined(SystemVerilogSubsetParser),
+)
+log.warning(
+    "Visitor module: %s @ %s",
+    SystemVerilogSubsetVisitor.__module__,
+    where_defined(SystemVerilogSubsetVisitor),
+)
+
 
 class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
     def __init__(self):
@@ -58,127 +77,33 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         self.module_map = {}
         self.output_signals = set()
         self.strict_identifiers = False
-        self.vector_widths: dict[str, tuple[int, int]] = {}
-        self.vector_ranges: dict[str, tuple[int, int]] = {}
         self.logger = logging.getLogger("logictree.SVToLogicTreeLowerer")
         # name -> (msb, lsb), e.g., { "s": (3, 0), "y": (1, 0) }
 
-    def bit_width(self, name: str) -> int | None:
-        t = self.vector_widths.get(name)
-        return (abs(t[0] - t[1]) + 1) if t else None
-
     def _labels_from_case_item(self, ctx) -> Tuple[List[LogicConst], bool]:
-    
+
         labels: List[LogicConst] = []
-    
+
         raw_text = ctx.getText()
         is_default = raw_text.strip().lower().startswith("default")
-    
+
         if is_default:
-            labels.append(LogicConst("default"))
-            return labels, True
-    
+            # No need to manufacture a bogus LogicConst just return
+            return [], True
+
         # Fallback: attempt to extract constant expressions
-        expr_ctxs = getattr(ctx, "constant_expression", []) or getattr(ctx, "expression", [])
+        expr_ctxs = getattr(ctx, "constant_expression", []) or getattr(
+            ctx, "expression", []
+        )
         for expr_ctx in expr_ctxs:
             expr_node = self.visit(expr_ctx)
             if not isinstance(expr_node, LogicConst):
-                raise ValueError(f"Expected LogicConst, got {type(expr_node)} from {expr_ctx.getText()}")
+                raise ValueError(
+                    f"Expected LogicConst, got {type(expr_node)} from {expr_ctx.getText()}"
+                )
             labels.append(expr_node)
-    
+
         return labels, False
-    #def _labels_from_case_item(self, ci) -> tuple[list[LogicConst], bool]:
-    #    """
-    #    Extracts labels from a CaseItem parse context.
-    #    Returns (labels, is_default).
-    #    - labels: list of LogicConst for case labels.
-    #    - is_default: True if this is a default case.
-    #    """
-    #    # 1) Check if it's explicitly marked default (for robust grammars)
-    #    if hasattr(ci, "DEFAULT") and callable(ci.DEFAULT) and ci.DEFAULT():
-    #        return [], True
-    #
-    #    labels: list[LogicConst] = []
-    #
-    #    # 2) Check for case_item_label_list → case_item_label
-    #    try:
-
-    #        if hasattr(ci, "case_item_label_list") and ci.case_item_label_list():
-    #            for lab in ci.case_item_label_list().case_item_label():
-    #                # Instead of parsing txt:
-    #                expr = lab.expression()
-    #                if expr:
-    #                    val = self.visit(expr)
-    #                    if isinstance(val, LogicConst):
-    #                        labels.append(val)
-    #            if labels:
-    #                return labels, False
-    #        #if hasattr(ci, "case_item_label_list") and ci.case_item_label_list():
-    #        #    for lab in ci.case_item_label_list().case_item_label():
-    #        #        txt = lab.getText()
-    #        #        try:
-    #        #            bv = BitVector.parse(txt)
-    #        #            labels.append(LogicConst(bv))
-    #        #            continue
-    #        #        except Exception:
-    #        #            # Fallback: try visiting the label
-    #        #            if hasattr(lab, "expression") and lab.expression():
-    #        #                expr_node = self.visit(lab.expression())
-    #        #                if isinstance(expr_node, LogicConst):
-    #        #                    labels.append(expr_node)
-    #        #    if labels:
-    #        #        return labels, False
-    #    except Exception:
-    #        pass  # fall through to next attempt
-    #
-    #    # 3) Alternative form: expression_list directly on the item
-    #    try:
-    #        if hasattr(ci, "expression_list") and ci.expression_list():
-    #            for e in ci.expression_list().expression():
-    #                txt = e.getText()
-    #                try:
-    #                    bv = BitVector.parse(txt)
-    #                    labels.append(LogicConst(bv))
-    #                except Exception:
-    #                    expr_node = self.visit(e)
-    #                    if isinstance(expr_node, LogicConst):
-    #                        labels.append(expr_node)
-    #            if labels:
-    #                return labels, False
-
-    #            if hasattr(ci, "expression_list") and ci.expression_list():
-    #                for lab in ci.expression_list().expression():
-    #                    # Instead of parsing txt:
-    #                    expr = lab.expression()
-    #                    if expr:
-    #                        val = self.visit(expr)
-    #                        if isinstance(val, LogicConst):
-    #                            labels.append(val)
-    #                if labels:
-    #                    return labels, False
-    #    except Exception:
-    #        pass
-    #
-    #    # 4) Final fallback: raw text parsing, check for default keyword
-    #    try:
-    #        raw = ci.getText()
-    #        head = raw.split(":", 1)[0].strip()
-    #        parts = [p.strip() for p in head.split(",") if p.strip()]
-    #        if head.lower() == "default":
-    #            return [LogicConst("default")], True
-    #        #if len(parts) == 1 and parts[0].lower() == "default":
-    #        #    return [LogicConst("default")], True
-    #        #for p in parts:
-    #        #    bv = BitVector.parse(p)
-    #        #    labels.append(LogicConst(bv))
-    #        #if labels:
-    #        #    return labels, False
-    #    except Exception:
-    #        pass
-    #
-    #    # 5) No labels and not default? Treat as empty (should be unreachable)
-    #    raise ValueError("Reached default fall through no labels and no default!")
-    #    return ["ERROR"], False
 
     def _parse_const(self, txt: str):
         """
@@ -190,16 +115,16 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         m = _BINARY_RE.match(txt)
         if m:
             width = int(m.group("width"))
-            base  = m.group("base").lower()
+            base = m.group("base").lower()
             digits = m.group("digits").replace("_", "")
-    
+
             if any(ch in "xXzZ" for ch in digits):
                 if not getattr(self, "allow_unknown_bits", False):
                     raise ValueError(f"Unknown bits in literal: {txt}")
                 # Coerce X/Z to 0 for now (documented behavior)
                 trans = str.maketrans({"x": "0", "X": "0", "z": "0", "Z": "0"})
                 digits = digits.translate(trans)
-    
+
             if base == "b":
                 val = int(digits, 2)
             elif base == "o":
@@ -211,26 +136,30 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             else:
                 raise ValueError(f"Unsupported base in literal: {txt}")
             return val, width
-    
+
         # plain decimal
         if txt.isdigit():
             return int(txt), None
-    
+
         raise ValueError(f"Unsupported literal syntax: {txt}")
 
     def visit(self, tree):
         # E.g., AndExprContext -> visitAndExpr
-        name = type(tree).__name__.replace('Context','')
-        meth = getattr(self, f'visit{name}', None)
+        name = type(tree).__name__.replace("Context", "")
+        meth = getattr(self, f"visit{name}", None)
         try:
             txt = tree.getText()
         except Exception:
             txt = "<no text>"
-        log.debug(f"[DISPATCH] {name} -> {'visit'+name if meth else 'visitChildren'} :: {txt}")
+        log.debug(
+            f"[DISPATCH] {name} -> {'visit'+name if meth else 'visitChildren'} :: {txt}"
+        )
         return meth(tree) if meth else self.visitChildren(tree)
 
     def lower(self, ast):
-        assert isinstance(ast, dict) and ast.get("modules"), "Expected a parsed AST with modules"
+        assert isinstance(ast, dict) and ast.get(
+            "modules"
+        ), "Expected a parsed AST with modules"
         mod = ast["modules"][0]  # just the first module for now
 
         if not mod["items"]:
@@ -259,7 +188,7 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
                 return ctx.Identifier().getText()
             elif hasattr(ctx, "identifier"):
                 return ctx.identifier().getText()
-            
+
             if hasattr(ctx, "blocking_assignment") and ctx.blocking_assignment():
                 return ctx.blocking_assignment().Identifier().getText()
 
@@ -286,16 +215,19 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         self.signal_map = {}
         self.output_signals = set()
 
-        #identifier_ctx = ctx.module_identifier()
-        module_name  = ctx.module_identifier().getText()
-        self.module_name  = module_name
+        # identifier_ctx = ctx.module_identifier()
+        module_name = ctx.module_identifier().getText()
+        self.module_name = module_name
         log.debug(f"Parsing module: {module_name}")
 
         mod_obj = Module(name=module_name)
         self.current_module = mod_obj
+        log.debug(f"type(mod_obj.signal_map): {type(mod_obj.signal_map)}")
+        log.debug(f"Module: {Module}")
+        log.debug(f"Module.__dataclass_fields__: {Module.__dataclass_fields__}")
 
         # when you create the module
-        self.current_module.vector_widths = {}   # name -> (msb:int, lsb:int)
+        self.current_module.vector_widths = {}  # name -> (msb:int, lsb:int)
 
         port_list_ctx = ctx.port_list()
         if port_list_ctx:
@@ -313,8 +245,8 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         log.debug(f">>>post visit: {ctx.getText()}")
 
         mod_obj.ports = ports.copy()
-        mod_obj.signal_map = self.current_module.signal_map.copy()
-
+        # mod_obj.signal_map = self.current_module.signal_map.copy()
+        mod_obj.signal_map.update(self.current_module.signal_map)
 
         # Output debug summaries
         log.debug("Signal map contents after visiting module:")
@@ -324,24 +256,25 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         log.debug("Output signals detected:")
         for out in self.output_signals:
             log.debug(f"  {out}")
-        
+
         self.module_map[module_name] = mod_obj
         log.debug(f"setting module_map[{module_name}] = {mod_obj}")
-        self.current_module = None #Clear after processing
+        log.debug("Module Dump:\n%s", pformat(mod_obj.__dict__, indent=2))
+        self.current_module = None  # Clear after processing
         return mod_obj
-
-
 
     def visitModule_item(self, ctx):
         log.debug("visitModule_item")
-        #log.debug(f"ctx.getChildren(): {ctx.getChildren()}")
+        # log.debug(f"ctx.getChildren(): {ctx.getChildren()}")
         for child in ctx.getChildren():
-            log.debug(f"Child of module_item: {type(child).__name__}, text: {child.getText()}")
+            log.debug(
+                f"Child of module_item: {type(child).__name__}, text: {child.getText()}"
+            )
             if isinstance(child, SystemVerilogSubsetParser.Continuous_assignContext):
                 log.debug("Detected Continuous_assignContext")
                 return self.visitContinuous_assign(child)
         return self.visitChildren(ctx)
-    
+
     def visitPort_list(self, ctx):
         log.debug("visitPort_list")
         for port_ctx in ctx.port():
@@ -352,50 +285,54 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         direction_tok = ctx.getChild(0).getText()
         if direction_tok not in ("input", "output", "inout"):
             return
-    
+
         text = ctx.getText()  # e.g. "inputlogic[1:0]s" or "outputlogicy"
         # parse optional packed range "[msb:lsb]"
         msb = lsb = None
         m = re.search(r"\[(\d+)\s*:\s*(\d+)\]", text)
         if m:
             msb, lsb = int(m.group(1)), int(m.group(2))
-    
+
         # collect identifiers (prefer tokens, fall back to a regex parse)
         try:
-            ids = [t.getText() for t in ctx.getTokens(SystemVerilogSubsetParser.Identifier)]
+            ids = [
+                t.getText() for t in ctx.getTokens(SystemVerilogSubsetParser.Identifier)
+            ]
         except Exception:
             ids = []
         if not ids:
             # fallback: strip up to the ']' if any, then strip type keywords, split on commas/space
-            tail = text.split(']', 1)[-1] if ']' in text else text
-            tail = re.sub(r'^(logic|wire|reg|signed|unsigned)+', '', tail)
-            ids = [tok for tok in re.split(r'[,\s]+', tail) if tok]
-    
+            tail = text.split("]", 1)[-1] if "]" in text else text
+            tail = re.sub(r"^(logic|wire|reg|signed|unsigned)+", "", tail)
+            ids = [tok for tok in re.split(r"[,\s]+", tail) if tok]
+
         for name in ids:
             # create/record the var
             if name not in self.current_module.signal_map:
                 var = self.current_module.signal_map.get(name, LogicVar(name))
                 width = 0
                 if msb is not None:
-                    width = abs(msb-lsb) + 1
-                    var = var.with_width(width)
+                    width = abs(msb - lsb) + 1
+                    # var = var.with_width(width)
+                    var = LogicVar(name, width=width)
                     self.current_module.vector_widths[name] = (msb, lsb)
                 else:
-                    var = var.with_width(1)
+                    # var = var.with_width(1)
+                    var = LogicVar(name, width=1)
 
                 self.current_module.signal_map[name] = var
-    
+
             # record vector width if present
             if msb is not None:
                 self.current_module.vector_widths[name] = (msb, lsb)
                 width_str = f"[{msb}:{lsb}]"
             else:
                 width_str = "scalar"
-    
+
             # track output ports (matches your earlier behavior of exposing only outputs)
             if direction_tok == "output" and name not in self.current_module.ports:
                 self.current_module.ports.append(name)
-    
+
             # debug like before
             self.logger.debug(f"Port {direction_tok:<6} {name:<10} width={width_str}")
 
@@ -454,20 +391,23 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             if isinstance(case_node, control.CaseStatement):
                 log.debug("Located a CaseStatement node!")
                 # Extract LHS from teh first case item (assumes consistemnt assignment target)
-                lhs = case_node.items[0].body.lhs
+                if case_node.items and case_node.items[0].body:
+                    lhs = case_node.items[0].body[0].lhs
                 self.current_module.signal_map[lhs] = case_node
-                log.debug(f"Registered logic for {lhs}:\n{pretty_print(self.current_module.signal_map[lhs])}")
-                #log.debug(pretty_print(case_node))
+                log.debug(
+                    f"Registered logic for {lhs}:\n{pretty_print(self.current_module.signal_map[lhs])}"
+                )
+                log.debug("case_node: %s", pretty_print(case_node))
             return case_node
 
         elif ctx.blocking_assignment():
             log.debug("visitStatement blocking_assigment")
             assign_ctx = ctx.blocking_assignment()
-            lhs_name   = assign_ctx.variable_lvalue().getText()
-            lhs        = lhs_name
-            rhs_expr   = assign_ctx.expression()
-            rhs_tree   = self.visit(rhs_expr)
-            assign_node  = control.LogicAssign(lhs=lhs, rhs=rhs_tree)
+            lhs_name = assign_ctx.variable_lvalue().getText()
+            lhs = lhs_name
+            rhs_expr = assign_ctx.expression()
+            rhs_tree = self.visit(rhs_expr)
+            assign_node = control.LogicAssign(lhs=lhs, rhs=rhs_tree)
             self.current_module.signal_map[lhs] = rhs_tree
             log.info(f"[statement assign] {assign_node}")
             return assign_node
@@ -485,15 +425,18 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         rhs_tree = self.visit(ctx.expression())
         lhs_var = self.current_module.signal_map.get(lhs, LogicVar(lhs))
         node = LogicAssign(lhs=lhs_var, rhs=rhs_tree)
+        log.debug(
+            f"Assigning signal_map.get() to current_module.assignments[{lhs}] = {node}"
+        )
         self.current_module.assignments[lhs] = node
         log.debug(f"[statement assign] {node}")
         return node
 
-    #def visitBlocking_assignment(self, ctx):
+    # def visitBlocking_assignment(self, ctx):
     #    log.debug("visitBlocking_assignment")
     #    lhs = ctx.variable_lvalue().getText()
     #    rhs_tree = self.visit(ctx.expression())
-    #    
+    #
     #    assign_node = control.LogicAssign(lhs=lhs, rhs=rhs_tree)
     #    self.current_module.signal_map[lhs] = rhs_tree
     #    log.debug(f"[statement assign] {assign_node}")
@@ -502,32 +445,43 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
     def visitIf_statement(self, ctx):
         log.debug("DEBUG: visitIf_statement()")
         cond_tree = self.visit(ctx.expression())
-    
+
         then_stmt_ctx = ctx.statement(0)
         else_stmt_ctx = ctx.statement(1) if ctx.ELSE() else None
-    
+
         then_result = self.visit(then_stmt_ctx)
-        if not isinstance(then_result, control.LogicAssign):
-            raise TypeError(f"Expected LogicAssign from then-branch, got {type(then_result)}")
+        if not isinstance(then_result, LogicAssign):
+            raise TypeError(
+                f"Expected LogicAssign from then-branch, got {type(then_result)}"
+            )
         lhs_then = then_result.lhs
         then_tree = then_result.rhs
-    
+
         if else_stmt_ctx:
             else_result = self.visit(else_stmt_ctx)
-            if not isinstance(else_result, control.LogicAssign):
-                raise TypeError(f"Expected LogicAssign from else-branch, got {type(else_result)}")
+            if not isinstance(else_result, LogicAssign):
+                raise TypeError(
+                    f"Expected LogicAssign from else-branch, got {type(else_result)}"
+                )
             lhs_else = else_result.lhs
             else_tree = else_result.rhs
         else:
             lhs_else = lhs_then
             else_tree = ops.LogicConst(0)
-    
+
         if lhs_then != lhs_else:
             raise NotImplementedError("Mismatched lhs in if/else assignment")
-    
-        mux_tree = ops.LogicOp("MUX", [cond_tree, then_tree, else_tree])
+
+        mux_tree = LogicMux(selector=cond_tree, if_true=then_tree, if_false=else_tree)
+        # assign = control.LogicAssign(lhs_then, mux_tree.to_primitives().simplify())
+        from logictree.transforms.simplify import simplify_logic_tree
+
+        assign = LogicAssign(lhs_then, simplify_logic_tree(mux_tree.to_primitives()))
+
         self.current_module.signal_map[lhs_then] = mux_tree
-        return control.LogicAssign(lhs_then, mux_tree)
+        self.current_module.assignments[lhs_then] = assign
+
+        return assign
 
     def visitExpression_list(self, ctx):
         log.debug("visitExpression_list")
@@ -538,39 +492,89 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         children = list(ctx.getChildren())
         log.debug(f"num_children: {len(children)}")
         log.debug(f"children: {children}")
-        try:
-            lhs_ctx = ctx.variable_lvalue()
-            rhs_ctx = ctx.expression()
-    
-            lhs = lhs_ctx.getText()
-            rhs_text = rhs_ctx.getText()
 
-            log.debug(f"lhs: {lhs}")
-            log.debug(f"rhs_text: {rhs_text}")
-    
-            # Resolve LHS var from the module's signal map
+        try:
+            # LHS
+            lhs_ctx = ctx.variable_lvalue()
+            lhs = lhs_ctx.getText()
+
+            for name, val in vars(self.current_module).items():
+                if isinstance(val, dataclasses.Field):
+                    log.warn(f"{name} is still a Field object: {val}")
+
+            log.debug(f"current_module: {self.current_module}")
+
+            log.debug("Module class: %s", type(self.current_module))
+            log.debug("Module.__module__: %s", type(self.current_module).__module__)
+            log.debug("Module.__dict__: %s", self.current_module.__dict__)
             lhs_var = self.current_module.signal_map.get(lhs, LogicVar(lhs))
 
-            # Build the RHS tree via vistor dispatch
-            rhs_tree = self.visit(rhs_ctx)
-    
+            log.debug(
+                f"type(self.current_module.assignments) = {type(self.current_module.assignments)}"
+            )
+            log.debug(
+                f"type(self.current_module.signal_map) = {type(self.current_module.signal_map)}"
+            )
+            # RHS is always child[3] in "assign <lhs> = <rhs> ;"
+            rhs_ctx = ctx.getChild(3)
+            rhs_text = rhs_ctx.getText()
+            log.debug(f"lhs: {lhs}, rhs_text: {rhs_text}")
+
+            rhs_tree = self.visit(rhs_ctx)  # must dispatch visitor!
+            log.debug(f"assign LHS = {lhs}, RHS tree = {rhs_tree}")
+
+            from logictree.utils.debug import assert_no_fields
+
             log.debug(f"Creating assign: {lhs_var} = {rhs_tree}")
             assign_node = LogicAssign(lhs=lhs_var, rhs=rhs_tree)
 
+            field_name, field_val = contains_field_object(assign_node)
+            if field_name:
+                log.error(
+                    f"assign_node.{field_name} is a dataclasses.Field: {field_val}"
+                )
+                raise TypeError(
+                    f"assign_node contains uninitialized dataclass field '{field_name}'"
+                )
+
+            log.debug("assign_node: %s", dataclasses.asdict(assign_node))
+
+            for attr_name, attr_val in vars(assign_node).items():
+                if isinstance(attr_val, dataclasses.Field):
+                    log.error(
+                        f" assign_node.{attr_name} is a dataclasses.Field: {attr_val}"
+                    )
+                else:
+                    log.debug(f" assign_node.{attr_name} = {attr_val}")
+
+            # optional viz label
             try:
                 from logictree.utils.display import pretty_inline
-                rhs_tree.set_viz_label(f"{lhs} = {pretty_inline(rhs_tree)}")
+                from logictree.utils.overlay import set_label
+
+                set_label(rhs_tree, f"{lhs} = {pretty_inline(rhs_tree)}")
+                # rhs_tree.set_viz_label(f"{lhs} = {pretty_inline(rhs_tree)}")
             except Exception as e:
                 log.debug("Could not set viz label: %s", e)
 
-            self.current_module.assignments[lhs] = assign_node
+            self.current_module.assignments[lhs_var.name] = assign_node
 
             log.debug("Assignments collected:")
             for k, assign in self.current_module.assignments.items():
                 log.debug(f"  {k} = {assign}")
 
+            assert_no_fields(assign_node, name="assign_tree")
             return assign_node
+
+        except AttributeError as e:
+            if "'Field' object has no attribute 'get'" in str(e):
+                log.error(
+                    "Likely dataclass mis-initialization: caught Field object in assign_node"
+                )
+                raise  # or continue
+            raise
         except Exception as e:
+            log.warning("visitContinuous_assign failed to parse assign")
             log.warning(f"type(ctx.getText()): {type(ctx.getText()).__name__}")
             log.warning(f"Failed to parse assign: {ctx.getText()} — {e}")
             return None
@@ -582,11 +586,12 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             # body: adjust to your rule name (statement / statement_or_null)
             body = self.visit(ci.statement())
             labels, is_default = self._labels_from_case_item(ci)
-            log.debug(f"visitCase_statement: labels, is_default: {labels}, {is_default}")
+            log.debug(
+                f"visitCase_statement: labels, is_default: {labels}, {is_default}"
+            )
             case_item = control.CaseItem(labels=labels, default=is_default, body=body)
             items.append(case_item)
         return control.CaseStatement(selector=selector_node, items=items)
-
 
     def visitExpression(self, ctx):
         log.debug("visitExpression fallback hit")
@@ -594,10 +599,10 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
 
     def _as_int_if_const(self, node):
         return node.value if isinstance(node, LogicConst) else node
-    
+
     def visitBitSelectExpr(self, ctx):
         base = self.visit(ctx.expression(0))
-        idx  = self.visit(ctx.expression(1))
+        idx = self.visit(ctx.expression(1))
         # accept int or small const wrappers with .value
         if not isinstance(idx, int):
             v = getattr(idx, "value", None)
@@ -608,13 +613,13 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             else:
                 raise ValueError("Only constant bit-select index supported for now")
         return BitSelect(base, idx)
-    
+
     def visitPartSelectExpr(self, ctx):
         # expression '[' expression ':' expression ']'
         base = self.visit(ctx.expression(0))
-        msb  = self.visit(ctx.expression(1))
-        lsb  = self.visit(ctx.expression(2))
-    
+        msb = self.visit(ctx.expression(1))
+        lsb = self.visit(ctx.expression(2))
+
         def _as_int(n):
             if isinstance(n, int):
                 return n
@@ -625,7 +630,7 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             if isinstance(v, str) and v.isdigit():
                 return int(v)
             raise ValueError("Only constant part-select bounds supported for now")
-    
+
         return PartSelect(base, _as_int(msb), _as_int(lsb))
 
     def visitConcatExpr(self, ctx):
@@ -673,94 +678,82 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
         return ops.XnorOp(lhs, rhs)
 
     def visitEqExpr(self, ctx):
-        kids = list(ctx.getChildren())  # <lhs> '==' <rhs>
-        lhs  = self.visit(kids[0])
-        rhs  = self.visit(kids[2])
-    
-        if self._is_logic_var(lhs) and self._is_const(rhs):
-            return self._expand_eq_var_const(self._coerce_var_node(lhs), self._to_int(rhs))
-        if self._is_logic_var(rhs) and self._is_const(lhs):
-            return self._expand_eq_var_const(self._coerce_var_node(rhs), self._to_int(lhs))
-    
-        raise NotImplementedError("Equality lowering supported for <vector> == <const> only")
+        lhs = self.visit(ctx.expression(0))
+        rhs = self.visit(ctx.expression(1))
 
-    def _is_logic_var(self, node):
-        # If IdExpr returns the actual var node
-        for v in self.current_module.signal_map.values():
-            # identity or same name → treat as var
-            if node is v or getattr(node, "name", None) == getattr(v, "name", None):
-                return True
-        # If IdExpr returns the name (str)
-        return isinstance(node, str) and node in self.current_module.signal_map
-    
-    def _coerce_var_node(self, node):
-        # ensure we hold the node object your other passes expect
-        if isinstance(node, str):
-            return self.current_module.signal_map[node]
-        return node
-    
-    def _is_logic_var(self, node):
-        return isinstance(node, LogicVar)
-    
-    def _is_const(self, node):
-        if isinstance(node, int): return True
-        return hasattr(node, "value") and isinstance(node.value, int)
-    
-    def _expand_eq_var_const(self, var_node, const_int: int):
-        # Get declared range for the vector; store this during port/signal creation.
-        # Expect a dict like: self.current_module.vector_ranges[name] = (msb, lsb)
-        msb, lsb = self.current_module.vector_ranges[var_node.name]
-        width = abs(msb - lsb) + 1
-    
-        # mask/align constant to declared width
-        const_int &= (1 << width) - 1
-    
-        # Build MSB-first terms
+        # Case: vector equality with constant literal
+        if isinstance(lhs, LogicVar) and isinstance(rhs, LogicConst):
+            width = getattr(lhs, "width", None)
+            if width is None or width == 0:
+                width = (
+                    rhs.width
+                    if hasattr(rhs, "width")
+                    else (int(rhs.value).bit_length() or 1)
+                )
+
+            hi, lo = width - 1, 0
+            return self._expand_vector_comparison(lhs, hi, lo, int(rhs.value), "==")
+
+        # Case: part-select equality with constant
+        if isinstance(lhs, PartSelect) and isinstance(rhs, LogicConst):
+            return self._expand_vector_comparison(
+                lhs.base, lhs.msb, lhs.lsb, int(rhs.value), "=="
+            )
+
+        # Fallback
+        return EqOp(lhs, rhs)
+
+    def visitNeqExpr(self, ctx):
+        lhs = self.visit(ctx.expression(0))
+        rhs = self.visit(ctx.expression(1))
+
+        if isinstance(rhs, LogicConst):
+            if isinstance(lhs, LogicVar):
+                width = getattr(lhs, "width", None)
+                if width is None or width == 0:
+                    width = (
+                        rhs.width
+                        if hasattr(rhs, "width")
+                        else (int(rhs.value).bit_length() or 1)
+                    )
+
+                hi, lo = width - 1, 0
+                return self._expand_vector_comparison(lhs, hi, lo, int(rhs.value), "!=")
+
+            if isinstance(lhs, PartSelect):
+                return self._expand_vector_comparison(
+                    lhs.base, lhs.msb, lhs.lsb, int(rhs.value), "!="
+                )
+
+        return NeqOp(lhs, rhs)
+
+    def _expand_vector_comparison(
+        self, base, hi: int, lo: int, const_val: int, op: str
+    ):
+        """
+        Expand (vector == const) or (vector != const) into bit-level EqOps joined by AndOps.
+        """
+        width = abs(hi - lo) + 1
+        const_val &= (1 << width) - 1
+
         terms = []
-        for pos in range(width):  # pos 0 = MSB
-            idx = (msb - pos) if msb >= lsb else (lsb + (width - 1 - pos))
-            sel = BitSelect(var_node, idx)
-            bit = (const_int >> (width - 1 - pos)) & 1
-            terms.append(sel if bit else ops.NotOp(sel))
-    
+        # Align bit 0 of const_val with LSB of the slice
+        for i in range(width):
+            bit_index = lo + i if lo <= hi else hi + i
+            bit_val = (const_val >> i) & 1
+            bit_node = BitSelect(base, bit_index)
+            terms.append(EqOp(bit_node, LogicConst(bit_val)))
+
         node = terms[0]
         for t in terms[1:]:
-            node = ops.AndOp(node, t)
-        return node
-    
-    def _to_int(self, c):
-        if isinstance(c, int):
-            return c
-        v = getattr(c, "value", None)
-        if isinstance(v, int):
-            return v
-        try:
-            return int(c)
-        except Exception:
-            pass
-        raise TypeError(f"Cannot convert const {c!r} to int")
-    
-    def _expand_eq_var_const(self, var_node, const_value):
-        const_int = self._to_int(const_value)  # <-- important
-    
-        left, right = self.current_module.vector_widths.get(var_node.name, (None, None))
-        if left is None or right is None:
-            # scalar case
-            return var_node if const_int else ops.NotOp(var_node)
-    
-        width = abs(left - right) + 1
-    
-        terms = []
-        for pos in range(width):                 # pos 0 = MSB, pos width-1 = LSB
-            idx = (left - pos) if left >= right else (left + (width - 1 - pos))
-            sel = BitSelect(var_node, idx)
-            bit = (const_int >> (width - 1 - pos)) & 1
-            terms.append(sel if bit else ops.NotOp(sel))
-    
-        node = terms[0]
-        for t in terms[1:]:
-            node = ops.AndOp(node, t)
-        return node
+            node = AndOp(node, t)
+
+        if op == "==":
+            return node
+        elif op == "!=":
+            return NotOp(node)
+        else:
+            raise ValueError(f"Unsupported comparison op: {op}")
 
     def visitParenExpr(self, ctx):
         log.info("visitParenExpr with: %s", ctx.getText())
@@ -768,19 +761,23 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
 
     def visitRange(self, ctx):
         log.debug("visitRange")
-        return self.visitChildren(ctx)
+        hi = int(ctx.DecimalNumber(0).getText())
+        lo = int(ctx.DecimalNumber(1).getText())
+        width = abs(hi - lo) + 1
+        return (hi, lo, width)
+        # return self.visitChildren(ctx)
 
     def _parse_binary_literal(self, txt: str):
         # e.g. "4'b1010" or "2'B01"
         width_str, rest = txt.split("'")
         width = int(width_str)
-        bits = rest[1:]                  # "1010"
+        bits = rest[1:]  # "1010"
         if any(ch in "xzXZ" for ch in bits):
             if not getattr(self, "allow_unknown_bits", False):
                 raise ValueError(f"Unknown bits in literal: {txt}")
-        val = int(bits.replace('_','').replace('x','0').replace('z','0'), 2)
+        val = int(bits.replace("_", "").replace("x", "0").replace("z", "0"), 2)
         return width, val
-    
+
     def visitLiteral(self, ctx):
         log.debug("visitLiteral")
         txt = ctx.getText()
@@ -793,29 +790,61 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             raise ValueError(f"Unsupported literal: {txt}")
 
     def visitConstExpr(self, ctx):
-        txt = ctx.getText().lower().replace("_", "")  # e.g., "4'b1001", "2'b10", "42"
-        if "'" not in txt:
-            # plain decimal
-            return int(txt)
-    
-        # based literal: <width>'<base><digits>
-        width_str, rest = txt.split("'", 1)
-        base_char = rest[0]            # b, d, h, o
-        digits    = rest[1:]           # e.g., "1001"
-    
-        base_map = {"b": 2, "d": 10, "h": 16, "o": 8}
-        base = base_map.get(base_char)
-        if base is None:
-            raise ValueError(f"Unsupported number base: {base_char}")
-    
-        # Optional: capture declared width if you care about zero-extension/truncation later
-        try:
-            self._last_const_width = int(width_str) if width_str else None
-        except ValueError:
-            self._last_const_width = None
-    
-        # NOTE: this assumes only 0/1 (no x/z) in tests
-        return int(digits, base)
+        log.debug("visitConstExpr")
+        text = ctx.getText()
+        log.debug(f"text: {text}")
+
+        # Binary, hex, decimal literals
+        if "'" in text:  # e.g. 2'b10, 4'hF, 8'd255
+            width_str, base_and_val = text.split("'")
+            log.debug(f"width_str: {width_str}, base_and_val: {base_and_val}")
+            width = int(width_str) if width_str else None
+            base = base_and_val[0].lower()
+            val_str = base_and_val[1:]
+            log.debug(f"width: {width}, base: {base}, val_str: {val_str}")
+
+            if base == "b":
+                value = int(val_str, 2)
+            elif base == "h":
+                value = int(val_str, 16)
+            elif base == "d":
+                value = int(val_str, 10)
+            else:
+                raise ValueError(f"Unsupported literal base: {base}")
+
+            log.debug(f"value: {value}")
+            const = LogicConst(value=value, width=width, base=base)
+            log.debug("Const constructed: %r (type=%s)", const, type(const.value))
+            return const
+
+        # Pure decimal (no base)
+        return LogicConst(value=int(text))
+
+    # def visitConstExpr(self, ctx):
+    #    tok = ctx.getText()
+    #
+    #    # Handle sized literals like 4'b1001, 8'hFF, etc.
+    #    if "'" in tok:
+    #        width_str, tail = tok.split("'", 1)
+    #        width = int(width_str)
+    #        base_char = tail[0].lower()
+    #        digits = tail[1:]
+    #
+    #        if base_char == "b":
+    #            value = int(digits, 2)
+    #        elif base_char == "h":
+    #            value = int(digits, 16)
+    #        elif base_char == "d":
+    #            value = int(digits, 10)
+    #        else:
+    #            raise ValueError(f"Unsupported literal base: {base_char}")
+    #
+    #        return LogicConst(value=value, width=width, base=base_char)
+    #
+    #    # Unsized decimal literal (e.g. "42")
+    #    else:
+    #        value = int(tok, 10)
+    #        return LogicConst(value=value)
 
     def visitIdExpr(self, ctx):
         log.debug("visitIdExpr")
@@ -828,4 +857,3 @@ class SVToLogicTreeLowerer(SystemVerilogSubsetVisitor):
             raise ValueError(f"Signal '{name}' not found in signal_map")
         log.debug("Implicit LogicVar for '%s'", name)
         return LogicVar(name)
-
