@@ -2,12 +2,11 @@ import os
 import subprocess
 from typing import Optional
 
+from logictree.utils.paths import OUTPUT_DIR
 from utils.graphviz_export import logic_tree_to_dot
 
-OUTPUT_DIR = os.path.join(os.path.dirname(__file__), "..", "output")
-os.makedirs(OUTPUT_DIR, exist_ok=True)
 
-def _run_dot(dot_path: str, fmt:str = "png", output_path: Optional[str] = None):
+def _run_dot(dot_path: str, fmt: str = "png", output_path: Optional[str] = None):
     """
     Internal helper to convert .dot file to the specified format using Graphviz.
 
@@ -24,7 +23,7 @@ def _run_dot(dot_path: str, fmt:str = "png", output_path: Optional[str] = None):
     if output_path is None:
         output_path = dot_path.replace(".dot", f".{fmt}")
 
-    if fmt not in ('svg', 'png'):
+    if fmt not in ("svg", "png"):
         raise ValueError(f"Unsupported output format: {fmt}")
 
     try:
@@ -38,13 +37,20 @@ def _run_dot(dot_path: str, fmt:str = "png", output_path: Optional[str] = None):
     except subprocess.CalledProcessError as e:
         raise RuntimeError(f"Graphviz failed: {e}")
 
-def to_svg(dot_path: str) -> str:
+
+def to_svg(tree, name="logic_tree") -> str:
+    dot = logic_tree_to_dot(tree)
+    dot_path = os.path.join(OUTPUT_DIR, f"output_sig_{name}.dot")
+    with open(dot_path, "w") as f:
+        f.write(dot.source if hasattr(dot, "source") else str(dot))
+    print(f"[DEBUG] Writing DOT file to {dot_path}")
     return _run_dot(dot_path, "svg")
 
-def to_png(tree, name="logic_tree") -> str:
-    dot_str = logic_tree_to_dot(tree)
-    dot_path = os.path.join(OUTPUT_DIR, f"{name}.dot")
-    with open(dot_path, "w") as f:
-        f.write(dot_str)
-    return _run_dot(dot_path, "png")
 
+def to_png(tree, name="logic_tree") -> str:
+    dot = logic_tree_to_dot(tree)
+    dot_path = os.path.join(OUTPUT_DIR, f"output_sig_{name}.dot")
+    with open(dot_path, "w") as f:
+        f.write(dot.source if hasattr(dot, "source") else str(dot))
+    print(f"[DEBUG] Writing DOT file to {dot_path}")
+    return _run_dot(dot_path, "png")
