@@ -1,2 +1,56 @@
-# nodes subpackage
-from . import base, control, hole, ops, types
+# src/logictree/nodes/__init__.py
+
+# --- Base ---
+from .base.base import LogicTreeNode
+
+# --- Control flow ---
+from .control.assign import LogicAssign
+from .control.case import CaseItem, CaseStatement
+from .control.ifstatement import IfStatement
+from .ops.comparison import EqOp, NeqOp
+from .ops.gates import AndOp, NandOp, NorOp, NotOp, OrOp, XnorOp, XorOp
+
+# --- Mux ---
+from .ops.mux import LogicMux
+
+# --- Core ops ---
+from .ops.ops import LogicConst, LogicOp, LogicVar
+
+# --- Selects ---
+from .selects import BitSelect, Concat, PartSelect
+
+# --- Structural ---
+from .struct.module import Module
+from .struct.statement import Statement
+
+__all__ = [
+    # Base
+    "LogicTreeNode",
+    # Ops
+    "LogicConst",
+    "LogicVar",
+    "LogicOp",
+    "AndOp",
+    "OrOp",
+    "NotOp",
+    "XorOp",
+    "NandOp",
+    "NorOp",
+    "XnorOp",
+    "EqOp",
+    "NeqOp",
+    # Selects
+    "BitSelect",
+    "PartSelect",
+    "Concat",
+    # Mux
+    "LogicMux",
+    # Control
+    "LogicAssign",
+    "IfStatement",
+    "CaseStatement",
+    "CaseItem",
+    # Structural
+    "Module",
+    "Statement",
+]
