@@ -1,7 +1,8 @@
 from abc import ABC, abstractmethod
-from typing import FrozenSet, Optional
+from dataclasses import dataclass, field
+from typing import FrozenSet, List, Optional
 
-from logictree.nodes.ops.ops import LogicVar
+from logictree.nodes import LogicVar
 
 
 class Statement(ABC):
@@ -36,3 +37,23 @@ class Statement(ABC):
     def writes_must(self) -> FrozenSet[LogicVar]:
         """Returns all LogicVars written to on all paths (must-write set)."""
         ...
+
+
+@dataclass(frozen=True)
+class BlockStatement(Statement):
+    """A sequence of multiple statements grouped as a single statement node."""
+
+    statements: List[Statement] = field(default_factory=list)
+
+    def free_vars(self) -> FrozenSet[LogicVar]:
+        return frozenset().union(*(s.free_vars() for s in self.statements))
+
+    def writes(self) -> FrozenSet[LogicVar]:
+        return frozenset().union(*(s.writes() for s in self.statements))
+
+    def writes_must(self) -> FrozenSet[LogicVar]:
+        # only vars written in *all* child statements
+        if not self.statements:
+            return frozenset()
+        must_sets = [s.writes_must() for s in self.statements]
+        return frozenset.intersection(*must_sets)
