@@ -1,15 +1,9 @@
-#GATE_TYPES = {
+# GATE_TYPES = {
 #    'AND', 'OR', 'NOT', 'NAND', 'NOR', 'XOR', 'XNOR',
 #    'MUX', 'EQ', 'IF', 'NEQ'
-#}
-GATE_TYPES = {
-    'AND', 'OR', 'NOT', 'NAND', 'NOR', 'XOR', 'XNOR',
-    'IF' 
-}
+# }
+GATE_TYPES = {"AND", "OR", "NOT", "NAND", "NOR", "XOR", "XNOR", "IF"}
 
-CONTROL_NODE_TYPES = {
-    'CaseStatement', 'AssignStatement', 'IfStatement'
-}
+CONTROL_NODE_TYPES = {"CaseStatement", "AssignStatement", "IfStatement"}
 
-COMMUTATIVE_OPS = {'AND', 'OR', 'XOR', 'XNOR'}
-
+COMMUTATIVE_OPS = {"AND", "OR", "XOR", "XNOR"}
