@@ -1,2 +1,20 @@
 # logictree utils __init__.py
-from .reduce import balanced_tree_reduce
+"""
+Utility subpackage for LogicTree.
+
+This collects utility functions that don’t belong in the core IR:
+- analysis
+- display
+- graphviz export
+- output helpers
+- reduction / simplification
+- serialization
+"""
+
+# from .overlay import overlay
+# from .analysis import *
+# from .display import *
+# from .graphviz_export import *
+# from .output import *
+# from .reduce import *
+# from .serialize import *
