@@ -4,53 +4,66 @@ from typing import Dict
 
 from dd.autoref import BDD
 
-from logictree.nodes.base import LogicTreeNode
-from logictree.nodes.ops import LogicOp
 from logictree.nodes.types import GATE_TYPES
 from logictree.utils.build import _build_bdd
 from logictree.utils.display import _pretty_print_expr, to_symbolic_expr_str
 
 log = logging.getLogger(__name__)
 
+
 def count_gate_type(tree, gate_name):
+    from logictree.nodes.base.base import LogicTreeNode
+    from logictree.nodes.ops.ops import LogicOp
+
     if isinstance(tree, LogicOp):
-        return int(tree.name == gate_name) + sum(count_gate_type(inp, gate_name) for inp in tree.inputs())
+        return int(tree.name == gate_name) + sum(
+            count_gate_type(inp, gate_name) for inp in tree.inputs()
+        )
     elif isinstance(tree, LogicTreeNode):
         return 0
     return 0
 
+
 # NEW: return the full breakdown as a dict
 def gate_breakdown(tree) -> Dict[str, int]:
     return {g: count_gate_type(tree, g) for g in GATE_TYPES}
+
 
 # CHANGED: return a scalar total (tests expect an int)
 def gate_count(tree) -> int:
     counts = gate_breakdown(tree)
     return sum(counts.values())
 
+
 def gate_summary(tree):
     counts = gate_breakdown(tree)
     return ", ".join(f"{k}:{v}" for k, v in counts.items() if v > 0)
 
+
 def _as_var_name(v) -> str:
     # Make sure BDD gets string variable names
     return getattr(v, "name", str(v))
+
 
 def get_logic_hash(tree, ordering=None, return_expr=False):
     bdd = BDD()
     var_map = {}
 
     # Collect inputs (prefer an explicit API if your nodes provide it)
-    inputs = tree.inputs() if hasattr(tree, "inputs") else (tree.children if hasattr(tree, "children") else [])
-    log.info("Collected inputs: %s", inputs)
+    # inputs = tree.inputs() if hasattr(tree, "inputs") else (tree.children if hasattr(tree, "children") else [])
+
+    from logictree.utils.traverse import collect_logic_vars
+
+    vars_ = sorted({v.name for v in collect_logic_vars(tree)})
+    log.info("Collected inputs: %s", vars_)
 
     # Declare BDD vars as strings
-    for var in inputs:
+    for var in vars_:
         bdd.declare(_as_var_name(var))
 
     node = _build_bdd(tree, bdd, var_map)
     expr = str(bdd.to_expr(node))
-    logic_hash = hashlib.sha256(expr.encode('utf-8')).hexdigest()
+    logic_hash = hashlib.sha256(expr.encode("utf-8")).hexdigest()
     expr_str = to_symbolic_expr_str(tree)
 
     if return_expr:
@@ -58,11 +71,14 @@ def get_logic_hash(tree, ordering=None, return_expr=False):
     else:
         return logic_hash
 
+
+# TODO: Update explain_logic_hash to use collect_logic_vars helper method to initialize inputs in BDD
 def explain_logic_hash(tree, ordering=None):
     bdd = BDD()
     var_map = {}
 
-    inputs = (tree.inputs() if hasattr(tree, "inputs") else [])
+    inputs = tree.inputs() if hasattr(tree, "inputs") else []
+    # vars_ = sorted({v.name for v in collect_logic_vars(tree)})
     if ordering is not None:
         inputs = ordering
     else:
@@ -74,7 +90,7 @@ def explain_logic_hash(tree, ordering=None):
 
     node = _build_bdd(tree, bdd, var_map)
     expr_str = str(bdd.to_expr(node))
-    hash_str = hashlib.sha256(expr_str.encode('utf-8')).hexdigest()
+    hash_str = hashlib.sha256(expr_str.encode("utf-8")).hexdigest()
     _pretty_print_expr(expr_str)
     log.info("\nSHA256 Logic Hash:\n:%s", hash_str)
     return expr_str, hash_str
