@@ -1,5 +1,5 @@
-#from logictree.nodes import LogicOp, LogicVar, LogicConst, LogicHole, LogicNode, CaseStatement, CaseItem, IfStatement, LogicAssign
-#import hashlib
+# from logictree.nodes import LogicOp, LogicVar, LogicConst, LogicHole, LogicNode, CaseStatement, CaseItem, IfStatement, LogicAssign
+# import hashlib
 import itertools
 
 from dd.autoref import BDD
@@ -10,33 +10,38 @@ from logictree.utils.build import _build_bdd
 
 
 # === LOGICTREE COMPARISON ===
-def compare_logic_trees(tree1, tree2, method='auto', debug=False):
-    if method == 'structure':
+def compare_logic_trees(tree1, tree2, method="auto", debug=False):
+    if method == "structure":
         return _compare_structure(tree1, tree2)
-    elif method == 'eval':
+    elif method == "eval":
         return _compare_truth_table(tree1, tree2)
-    elif method == 'bdd':
+    elif method == "bdd":
         return _compare_bdd(tree1, tree2)
-    elif method == 'hash':
+    elif method == "hash":
         return get_logic_hash(tree1) == get_logic_hash(tree2)
-    elif method == 'auto':
+    elif method == "auto":
         if _compare_structure(tree1, tree2):
-            if debug: print("Structure matched.")
+            if debug:
+                print("Structure matched.")
             return True
         if get_logic_hash(tree1) == get_logic_hash(tree2):
-            if debug: print("Hash matched.")
+            if debug:
+                print("Hash matched.")
             return True
         if _compare_bdd(tree1, tree2):
-            if debug: print("BDD matched.")
+            if debug:
+                print("BDD matched.")
             return True
         return False
     else:
         raise ValueError(f"Unknown comparison method: {method}")
 
+
 def _compare_structure(t1, t2):
     if t1.op != t2.op or len(t1.children) != len(t2.children):
         return False
     return all(_compare_structure(c1, c2) for c1, c2 in zip(t1.children, t2.children))
+
 
 def _compare_truth_table(t1, t2):
     inputs = sorted(set(t1.inputs()) | set(t2.inputs()))
@@ -46,12 +51,14 @@ def _compare_truth_table(t1, t2):
             return False
     return True
 
+
 def _compare_bdd(t1, t2):
     return to_bdd(t1) == to_bdd(t2)
 
+
 # === BDD BACKEND ===
 def to_bdd(tree: base.LogicTreeNode, ordering=None) -> int:
-    #bdd = _bdd.BDD()
+    # bdd = _bdd.BDD()
     bdd = BDD()
     var_map = {}
     inputs = sorted(tree.operands) if ordering is None else ordering
@@ -59,4 +66,3 @@ def to_bdd(tree: base.LogicTreeNode, ordering=None) -> int:
         bdd.declare(var)
     root = _build_bdd(tree, bdd, var_map)
     return bdd, root
-
