@@ -3,6 +3,7 @@ import pytest
 pytestmark = [pytest.mark.unit]
 
 from logictree.nodes.ops.ops import LogicTreeNode
+from logictree.transforms.simplify import simplify
 from tests.utils import EXCLUDED_CLASSES, all_subclasses, safe_instantiate
 
 
@@ -16,8 +17,8 @@ def test_all_gates_implement_simplify():
         if instance is None:
             continue
         try:
-            _ = instance.simplify()
+            _ = simplify(instance)
         except Exception:
             bad_simplify.append(cls.__name__)
 
-    assert not bad_simplify, f"Failed .simplify() on: {', '.join(bad_simplify)}"
+    assert not bad_simplify, f"Failed simplify() on: {', '.join(bad_simplify)}"
